rename files from shell to shells
