+++ src/patterns/index.js
@@ -8,7 +8,7 @@ export * as EmptyState  from './EmptyState';
 export * as ActionPanels  from './ActionPanels';
 export * as Feeds  from './Feed';
 export * as Forms  from './Form';
-export * as Shell  from './Shell';
+export * as Shells  from './Shells';
 export * as SignIn  from './SignIn';
 export * as Stats  from './Stats';
 export * as GridLists  from './GridLists';
